% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTriangleMatrix.R
\name{plotTriangleMatrix}
\alias{plotTriangleMatrix}
\title{plotTriangleMatrix plots two matrices z1 and z2 of equal size 
as a matrix plot such that the corresponding matrix elements 
are show in the same picture cell by diving the cell to 
two triangles.}
\usage{
plotTriangleMatrix(z1, z2, keep.rect = T, colRamp = NULL, xlab = NULL,
  ylab = NULL, main = NULL, xOffset = 0.03, yOffset = 0.03,
  cex.main = 1, cex.x = 1, cex.y = 1, ...)
}
\arguments{
\item{z1}{}

\item{z2}{}

\item{keep.rect}{If \code(true), individual elements in the 
figure are kept rectangular. If \code(false), the whole figure 
area is kept rectangular.}

\item{colRamp}{User-specified colorRamp function. By default, 
\code(grey()) is used.}

\item{xlab}{Labels in x-direction}

\item{ylab}{Labels in y-direction}

\item{main}{Main title}

\item{xOffset}{An offset value for x labels. By default, 
value 0.03 is used.}

\item{yOffset}{An offset value for y labels. By default, 
value 0.03 is used.}

\item{cex.main}{Character expasion value for the main title.}

\item{cex.x}{Character expasion value for x labels.}

\item{cex.y}{Character expasion value for y labels .}

\item{...}{Not implemented yet.}
}
\description{
plotTriangleMatrix plots two matrices z1 and z2 of equal size 
as a matrix plot such that the corresponding matrix elements 
are show in the same picture cell by diving the cell to 
two triangles.
}
